backend/requirements.txt
# Core web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# LLM — new google-genai SDK (Gemini 3)
google-genai>=1.0.0

# Agent orchestration
langgraph>=0.4.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Web scraping
firecrawl-py>=1.0.0
playwright>=1.48.0

# HTML parsing and DOM manipulation
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
pillow>=10.0.0
httpx>=0.27.0

# Cloud storage
google-cloud-storage>=2.18.0

# Observability (optional — set LANGCHAIN_TRACING_V2=true to enable)
langsmith>=0.1.0
